resolve multipledescribes spec
